Add tests for 'globus api' and fix var binding bug
`for x in ...` binds `x` dynamically inside the loop scope, and we
wanted to have a closure over `x`. An explicit function to act as a
closure avoids this sabotaging the `globus api` command.

This was discovered with the addition of a variety of unit tests and a
functional test for the `globus api` command.