mypy --disallow-untyped-defs on top-level modules
The modules in the top-level source dir are now all listed in mypy.ini
with disallow_untyped_defs=true

Various tweaks were needed to get mypy passing. The only significant
change to runtime logic was in the filter_fields helper method. Rather
than assuming that the inputs are always tuples, it now handles a
FormatField object gracefully (but does not check it against the
filter).