Add --allow-errors, --allow-redirects to 'api'
Allow error responses to be rendered verbatim (without our handlers
running) if `--allow-errors` is used. Do not follow redirects unless
`--allow-redirects/--location/-L` is used.

The new attributes of response and error classes make it easy to
unify the printing logic such that errors and responses are printed in
the same way. The only difference between the two is the name of the
text body attribute.