Add behavior like no-args-is-help to commands
Unlike the builtin no_args_is_help option, we do not want this
behavior to trigger on commands which take no arguments (all args are
optional).

In our custom command class, override the behavior of `parse_args` to
check for MissingParameter errors combined with an empty arg list.
Such a situation indicates that a command requires arguments (e.g.
`globus transfer`) and none were given. In this situation, print the
MissingParameter error message, followed by the full helptext for the
current command, then exit(2) .

If there were arguments and a MissingParameter error is encountered,
just reraise and trust click to handle it as normal.